# Rework plugins system and speed up rubygems