@deivid-rodriguez 2020-01-30
I updated the PR description, and the implementation should be ready.

I'm still thinking of adding some higher level tests, but I'm not sure whether to add them to `test/rubygems/test_gem_commands_install_command.rb` and `test/rubygems/test_gem_commands_uninstall_command.rb`, or to add a realworld tests similar to the one I've been running locally:

```bash
rm -f ~/.rbenv/versions/2.7.0/lib/ruby/gems/2.7.0/plugins/yard_plugin.rb && rake install && gem install yard:0.9.23 && cat ~/.rbenv/versions/2.7.0/lib/ruby/gems/2.7.0/plugins/yard_plugin.rb && gem install yard:0.9.24 && cat ~/.rbenv/versions/2.7.0/lib/ruby/gems/2.7.0/plugins/yard_plugin.rb && gem uninstall yard:0.24.0 --executables --backtrace && cat ~/.rbenv/versions/2.7.0/lib/ruby/gems/2.7.0/plugins/yard_plugin.rb && gem uninstall yard:0.23.0 --executables --backtrace && ls ~/.rbenv/versions/2.7.0/lib/ruby/gems/2.7.0/plugins/ -la
```

Essentially, install and uninstall gems with plugins and make sure the plugins location is properly updated.