@deivid-rodriguez 2020-01-30
Oh, also something to be noticed, if you try this locally and don't see a great improvement, it might be because you have etiher the `executable-hooks` or the `gem-wrappers` gems installed. These gems have rubygems plugins that are quite slow, so that's the bottleneck in that case, so the improvement is not apparent. But if you have `yard` installed, which also includes a rubygems plugin, it should still be much faster.