@deivid-rodriguez 2020-01-31
Ok, so I found and fixed some of bugs, and added a bunch of tests, so I'm happy with this now! Removing the WIP tag :)